cuda_kernel_pipeline/requirements.txt
PyGithub>=1.55
requests>=2.25.0
tqdm>=4.62.0
